lets start some better architecture
