+++ .vscode/c_cpp_properties.json
@@ -8,44 +8,48 @@
         {
             "name": "PlatformIO",
             "includePath": [
-                "/home/jlaustill/code/opcm/include",
-                "/home/jlaustill/code/opcm/src",
-                "/home/jlaustill/code/opcm/lib/TempSensor",
-                "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/Adafruit MAX31855 library",
-                "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/Adafruit BusIO",
-                "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/Wire/src",
-                "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/TimerFour",
-                "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/TimerFive",
-                "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/CAN_BUS_Shield/src",
+                "/home/jlaustill/repo/opcm/include",
+                "/home/jlaustill/repo/opcm/src",
+                "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/EEPROM/src",
+                "/home/jlaustill/repo/opcm/lib/TempSensor",
+                "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/Adafruit MAX31855 library",
+                "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/Adafruit BusIO",
+                "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/TimerFour",
+                "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/TimerFive",
+                "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/CAN_BUS_Shield/src",
                 "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/SPI/src",
                 "/home/jlaustill/.platformio/packages/framework-arduino-avr/cores/arduino",
                 "/home/jlaustill/.platformio/packages/framework-arduino-avr/variants/mega",
                 "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/SoftwareSerial/src",
-                "/home/jlaustill/code/opcm/lib/CircularList",
-                "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/LiquidCrystal",
-                "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/EEPROM/src",
+                "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/Wire/src",
+                "/home/jlaustill/repo/opcm/lib/CircularList",
+                "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/LiquidCrystal",
+                "/home/jlaustill/.platformio/lib/TimerOne",
+                "/home/jlaustill/.platformio/lib/ToneLibrary/src",
                 "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/HID/src",
                 ""
             ],
             "browse": {
                 "limitSymbolsToIncludedHeaders": true,
                 "path": [
-                    "/home/jlaustill/code/opcm/include",
-                    "/home/jlaustill/code/opcm/src",
-                    "/home/jlaustill/code/opcm/lib/TempSensor",
-                    "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/Adafruit MAX31855 library",
-                    "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/Adafruit BusIO",
-                    "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/Wire/src",
-                    "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/TimerFour",
-                    "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/TimerFive",
-                    "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/CAN_BUS_Shield/src",
+                    "/home/jlaustill/repo/opcm/include",
+                    "/home/jlaustill/repo/opcm/src",
+                    "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/EEPROM/src",
+                    "/home/jlaustill/repo/opcm/lib/TempSensor",
+                    "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/Adafruit MAX31855 library",
+                    "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/Adafruit BusIO",
+                    "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/TimerFour",
+                    "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/TimerFive",
+                    "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/CAN_BUS_Shield/src",
                     "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/SPI/src",
                     "/home/jlaustill/.platformio/packages/framework-arduino-avr/cores/arduino",
                     "/home/jlaustill/.platformio/packages/framework-arduino-avr/variants/mega",
                     "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/SoftwareSerial/src",
-                    "/home/jlaustill/code/opcm/lib/CircularList",
-                    "/home/jlaustill/code/opcm/.pio/libdeps/megaatmega2560/LiquidCrystal",
-                    "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/EEPROM/src",
+                    "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/Wire/src",
+                    "/home/jlaustill/repo/opcm/lib/CircularList",
+                    "/home/jlaustill/repo/opcm/.pio/libdeps/megaatmega2560/LiquidCrystal",
+                    "/home/jlaustill/.platformio/lib/TimerOne",
+                    "/home/jlaustill/.platformio/lib/ToneLibrary/src",
                     "/home/jlaustill/.platformio/packages/framework-arduino-avr/libraries/HID/src",
                     ""
                 ]
